Revert "Bump to dotnet/installer/main@d25a3bb 8.0.100-preview.2.23105.6 (#7769)"
Context: https://github.com/dotnet/runtime/issues/82597

This reverts commit 6cd0d38989178e1c3e5e1d71505b4aef26a1739b.

We need to do it because a recent update to the `Microsoft.NET.ILLink`
package broke our build:

   Microsoft.Android.Sdk.ILLink.csproj : error NU1202: Package Microsoft.NET.ILLink 8.0.0-preview.2.23123.5 is not compatible with net7.0 (.NETCoreApp,Version=v7.0). Package Microsoft.NET.ILLink 8.0.0-preview.2.23123.5 supports: net8.0 (.NETCoreApp,Version=v8.0)

This happened because the version of `Microsoft.NET.ILLink` that
we have configured in `en/Version.Details.xml` was removed from
the dotnet8 feed and the closest matching version is installed instead.

An attempt to fix it was to switch the `Microsoft.Android.Sdk.ILLink`
project back to `net8.0` but, alas, the new `Microsoft.NET.ILLink` package
version misses dependency on Cecil, also causing our build to fail:

   src/Microsoft.Android.Sdk.ILLink/ApplyPreserveAttribute.cs(9,12): error CS0234: The type or namespace name 'Cecil' does not exist in the namespace 'Mono'
   src/Xamarin.Android.Build.Tasks/Linker/MonoDroid.Tuner/LinkDesignerBase.cs(161,41): error CS0246: The type or namespace name 'MethodDefinition' could not be found
   ...
    626 Warning(s)
    359 Error(s)

Reverting 6cd0d38989178e1c3e5e1d71505b4aef26a1739b puts as back on the `alpha1`
release of net8 and fixes the build until we have a real solution available.