[build] remove darc dependency for System.IO.Hashing (#7855)
Maestro is currently failing to open new PRs due to:

    > darc update-dependencies --id 169161
    Looking up build with BAR id 169161
    Updating 'Microsoft.Dotnet.Sdk.Internal': '8.0.100-preview.3.23128.1' => '8.0.100-preview.3.23153.7' (from build '20230303.7' of 'https://github.com/dotnet/installer')
    Checking for coherency updates...
    Using 'Strict' coherency mode. If this fails, a second attempt utilizing 'Legacy' Coherency mode will be made.
    Coherency updates failed for the following dependencies:
    Unable to update System.IO.Hashing to have coherency with Microsoft.Dotnet.Sdk.Internal: https://github.com/dotnet/installer @ 9cf0095d0892d385a7e26772752d75e3bb68d8e4 does not contain dependency System.IO.Hashing

In 5d68699, we added a new dependency of System.IO.Hashing from
dotnet/runtime.

But for this to work, we'd have to:

1. List `System.IO.Hashing` as a dependency of dotnet/installer

or

2. Just use manage `$(SystemIOHashingPackageVersion)` to match the
   overall version of dotnet/runtime.

In dotnet/maui, at lot of times they've been going with no. 2. If
needed on release branches, we can also hardcode the number or switch
over to using a stable version on NuGet.org.

Let's remove the Maestro dependency and use
`$(MicrosoftNETCoreAppRefPackageVersion)` for now.