Fix unit test build failure
Commit 78ef0eaf didn't work; various unit tests failed with:

	error XAGJRNC7023: System.IO.DirectoryNotFoundException: Could not find a part of the path '…/tests/Mono.Android-Tests/Runtime-Microsoft.Android.Sdk/xa-internal/xa-remap-members.xml'.

Meaning explicitly mentioning the `_ConvertAndroidMamMappingFileToXml`
and `_CollectAndroidRemapMembers` targets was important!

A (slightly!) more in depth investigation suggests that the only
target that depends on the outputs of these targets is the
`_GenerateAndroidRemapNativeCode` target.  On the hope that this is
in fact the case, add a new `$(_GenerateAndroidRemapNativeCodeDependsOn)`
MSbuild property which is the target dependencies of the
`_GenerateAndroidRemapNativeCode` target, and add
`_ConvertAndroidMamMappingFileToXml` and `_CollectAndroidRemapMembers`
to `$(_GenerateAndroidRemapNativeCodeDependsOn)`.  This *should* ensure
that the targets are executed *as late as possible*.