Fix native code generation when marshal methods are disabled
Disabling marshal method generation turns off a lot of native code
generation in the `MarshalMethodsNativeAssemblyGenerator` class, but the
class is also responsible for outputting correctly sized cache area (an
array of X pointers) to be used by the native runtime to cache pointers
to `MonoImage` instances.

Native runtime trusts the managed runtime to generate correct code and,
thus, does not verify the size of generated array.  This trust,
unfortunately, was broken because with marshal methods disabled, the
native code generator output cache array that was 0 entries in size,
thus leading to segfault when attempting to run the application.

Fix the issue and also parametrize one of the on-device tests to be
built twice, with marshal methods explicitly disabled and explicitly
enabled.