[Xamarin.Android.Build.Tasks] Updating NuGet packages does not update the app (#7892)
Fixes https://github.com/xamarin/xamarin-android/issues/7890

We found and issue where if you upgraded a NuGet Package it did not up
date the assemblies in the `$(IntermediateOutputPath)android\assets` folder.
We have a unit test which checks this particular scenario, but it was
disabled on .net during the port to .net 6. We never got back to enable it.

The issue it turns out is that NuGet not longer populates the `ProjectLockFile`
property. As a result we never stored the `_NuGetAssetsTimestamp` in the
`build.props` file. This causes the `_CleanIntermediateIfNeeded` target
to be skipped when it really should run.

Switching over to using the `ProjectAssetsFile` property to populate the
`_NuGetAssetsTimestamp` property fixes the issue. Also the unit test was
updated to work on both Legacy and .net.