[Xamarin.Android.Build.Tasks] Fix AndroidPackagingOptionsExclude
Fixes https://github.com/xamarin/xamarin-android/issues/7902

Commit 2726a386 introduced the ability to ignore certain file patterns
when adding files to an apk. Unfortunately it had a bug in the code
and the test. The code would log a warning that is was ignoring the file,
but then add the file anyway. The test was looking for the wrong path
in the apk , so it would always pass the test.

This commit fixes both of these issues.