[tests] Add backup ssl sites in case of 429 response.
Our SSL test can fail quite often with a 429 Too Many Requests error.
This makes the CI very unstable as we are constantly having to wait and
retry the tests. So lets put that logic into the test itself. If we get
a 429 we should try some other ssl site.