Eliminate the need to use NDK when enabling marshal methods tracing
At XA build time we find and package the compiler's builtins libraries
for all the ABIs we support. This allows us to link libxamarin-app.so
without requiring NDK to be present on end user's machine.

Additionally, traces have now adjusted addresses and offsets for all
frames, to match bionic behavior.

Finally, the entire trace output is logged in a single message. This
makes everything faster, but also makes sure that the trace won't be
interleaved with unrelated messages in logcat.