Add Nightly Tests for Humanizer
Tests #7622

Localisation assemblies were not being fast deployed. Commit xamarin/monodroid@1f52d58
fixed that. However we need to add a unit test to make sure that these files are deployed and
actually produce the required results.

So lets use Humanizer to handle the localisation. This PR adds a test which checks that the
app on the device outputs the expected text which Humanizer is supposed to generate.
It also updates existing packaging tests to make sure that the required satellite assemblies are
packaged.