Fix integer to string conversion when generating code
Fixes: https://github.com/xamarin/xamarin-android/issues/7939
Context: https://github.com/dotnet/runtime/issues/13363
Context: https://www.unicode.org/charts/PDF/U2200.pdf (page 2)

LLVM IR generator outputs a number of integer values, relying on the
standard `ToString()` method to convert the value to string.  However,
since NET6, this conversion is culture-sensitive and in certain
cultures (list below) it will output the minus sign not as the standard
ASCII 0x2D character but rather as the Unicode 0x2212 (mathematical
operator "minus") character.  This breaks LLVM LLC:

    llc: environment.arm64-v8a.ll:554:7: error: expected value token
      stderr |                 i32 −1, ; apk_fd

Fix the problem by using invariant culture when converting **all**
integers and unknown objects to strings.  The reason all of them are
converted in this way is to avoid future changes to ICU and/or .NET
to-string conversion that might affect the resulting integer format.

Locales/cultures affected by this issue are as follows:

  * et
  * et-EE
  * eu
  * eu-ES
  * fi
  * fi-FI
  * fo
  * fo-DK
  * fo-FO
  * gsw
  * gsw-CH
  * gsw-FR
  * gsw-LI
  * hr
  * hr-BA
  * hr-HR
  * ksh
  * ksh-DE
  * lt
  * lt-LT
  * nb
  * nb-NO
  * nb-SJ
  * nn
  * nn-NO
  * no
  * rm
  * rm-CH
  * se
  * se-FI
  * se-NO
  * se-SE
  * sl
  * sl-SI
  * sv
  * sv-AX
  * sv-FI
  * sv-SE