[ci] Stop building classic test suites. (#7938)
Now that we no longer run any Classic test suites in main, we no longer need to build them with Xamarin.Android-Tests.sln.

The .NET test suites that we do use are built in Xamarin.Android.sln or built locally on the test agents with the installed XA test packages.

This saves about 7.5 minutes from the Mac Build.