Day spent debugging
It appears that the following might be cause of the "hang":

    04-07 19:39:36.121  4859  4859 I chromium: [INFO:CONSOLE(2)] "Uncaught ReferenceError: Blazor is not defined", source: https://0.0.0.0/ (2)

It is probably caused by this JavaScript fragment from Blazor:

    Blazor.start();
    window.__BlazorStarted = true;

And the Blazor class might not be initialized because of a problem with
delegates using the "old" registration mechanism when marshal methods
are enabled (there are quite a few of them in the test app).

lldb didn't reveal anything interesting, process state appeared to be
fine.