Convert `/tools` and `/build-tools` projects from `net472` to `$(DotNetStableTargetFramework)` (#7943)
Many of our projects in `/tools` and `/build-tools` target `net472`.  Update them to instead target `$(DotNetStableTargetFramework)`.

A side-effect of this is that there is no longer an `.exe` built, so we also need to update everything in our build system that calls these tools to use `dotnet foo.dll` instead of `mono foo.exe`.