[Xamarin.Android.Build.Tasks] Bump ZipFlushFilesLimit
Context https://i.azdo.io/1782014

On Windows customers are seeing the following error

```
Error XABLD7000: Xamarin.Tools.Zip.ZipException: Renaming temporary file failed: Permission denied
 at Xamarin.Tools.Zip.ZipArchive.Close() in /Users/runner/work/1/s/LibZipSharp/Xamarin.Tools.Zip/ZipArchive.cs:line 939
```
We have traced this to the use of the `MoveFileExW`/`MoveFileExA` API.
When libzip is trying to move its temp file to the final location,
Windows is raising this error.

```
Renaming temporary file failed: Permission denied
```

Turning off Anti Virus seems to help, however adding an
exclusion does not. This is very confusing. So we are unsure why this
error is being raised. The process has the correct permissions and the
file being moved is in the same directory, so its not a TEMP folder issue.

So perhaps its the number of temp files we create? Part of the `BuildApk`
system is that as we add files we `Flush` the zip file to commit those
changes to disk. This is partly to work around how libzip works. It does
not write any data to the main file until `zip_close` is called. So to
work around issues around too many files being open [1], we added this flush.

The limit of 50 files was picked out of a hat. So lets try pushing the
limit up a bit to see if that helps.

[1] https://github.com/xamarin/xamarin-android/commit/9166e0363417d6d121de37a765db8f2ba7cece7b