[ci] Automatically retry failed apk-instrumentation tests. (#7963)
The `apk-instrumentation` tests are a good candidate for using `retryCountOnTaskFailure`.  This is because they only take a few minutes to run, and have a low enough flaky rate that it is likely that rerunning the entire test suite will succeed.

As noted in https://github.com/xamarin/java.interop/pull/1095, AzDO's support for this isn't perfect.  The initial failure will still show up in the Error Log, however the task will be green and will not show up as a failed build.