[Xamarin.Android.Build.Tasks] Allow override of `uncompressedGlob`.
Users reported that when they added a custom `uncompressedGlob` entry to
the `$(AndroidBundleConfigurationFile)`, it was ignored.
This is because the `MergeArrayHandling` option we were using in
the `BuildAppBundle` task was set to `MergeArrayHandling.Replace`.

As a result we would just overwrite the user config with our own.
We should be using the `MergeArrayHandling.Union` option instead. This
with merge the arrays , but will remove duplicates. This should allow
users to provide additional `uncompressedGlob` options.

A unit test has been added to test this new behaviour.