[Xamarin.Android.Build.Tasks] stop redistributing `apksigner` (#7984)
Fixes: https://github.com/xamarin/xamarin-android/issues/6307

This reverts c50df1c5.

In the past, we had to redistribute `apksigner.jar`, because it
required JDK 11. This happened for build-tools 30.0.0 and was fixed in
built-tools 30.0.3. At the time we were nowhere close able to use JDK
11, and so we built it ourselves for JDK 1.8 and redistributed it.

However, .NET 6+ now *requires* JDK 11, because targets API-31 and
higher. "Classic" Xamarin.Android requires JDK 11 if you are targeting
API-31.

We can remove `apksigner` from `main` going forward in .NET 8.

Other changes:

* Update `XASdkTests.cs` to account for `*.apk.sig` files.