[ci] Bump NUnit versions (#8021)
With the new auto-retry for failed emulator tests (bff72425) we are
seeing some weird results.  For example:

The first run failed with:

	Failed DotNetInstallAndRun(True,True,"net8.0-android") [12 m 11 s]
	  Error Message:
	     `dotnet run` should succeed
	  Expected: True
	  But was:  False

	  Stack Trace:
	     at Xamarin.Android.Build.Tests.XASdkDeployTests.DotNetInstallAndRun(Boolean isRelease, Boolean xamarinForms, String targetFramework) in /Users/builder/azdo/_work/1/s/xamarin-android/tests/MSBuildDeviceIntegration/Tests/XASdkDeployTests.cs:line 87
	     at Xamarin.Android.Build.Tests.XASdkDeployTests.DotNetInstallAndRun(Boolean isRelease, Boolean xamarinForms, String targetFramework) in /Users/builder/azdo/_work/1/s/xamarin-android/tests/MSBuildDeviceIntegration/Tests/XASdkDeployTests.cs:line 87

The retry failed with:

	Failed DotNetInstallAndRun(True,True,"net8.0-android") [1 s]
	Error Message:
	 System.InvalidOperationException : Unknown abi: 
	Stack Trace:
	   at Xamarin.ProjectTools.AbiUtils.AbiToRuntimeIdentifier(String androidAbi) in /Users/builder/azdo/_work/1/s/xamarin-android/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Utilities/AbiUtils.cs:line 18
	 at Xamarin.ProjectTools.ProjectExtensions.SetRuntimeIdentifier(IShortFormProject project, String androidAbi) in /Users/builder/azdo/_work/1/s/xamarin-android/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Utilities/ProjectExtensions.cs:line 34
	 at Xamarin.Android.Build.Tests.XASdkDeployTests.DotNetInstallAndRun(Boolean isRelease, Boolean xamarinForms, String targetFramework) in /Users/builder/azdo/_work/1/s/xamarin-android/tests/MSBuildDeviceIntegration/Tests/XASdkDeployTests.cs:line 76
	 at System.RuntimeMethodHandle.InvokeMethod(Object target, Void** arguments, Signature sig, Boolean isConstructor)
	 at System.Reflection.MethodInvoker.Invoke(Object obj, IntPtr* args, BindingFlags invokeAttr)

There are other instances of tests failing with this message,
including timing tests that initially failed simply because the
process took too long to finish.

It seems like the valid ABIs aren't getting set on the second run,
which is done via `[OneTimeSetUp]`.  As crazy as this seems, since
the second run is a completely new process, this does appear to have
been [a bug in NUnit at some point][0].

I am not sure if all of our NUnit assemblies have been updated enough
to contain the fix for this.  Since it's a good idea to update to the
latest anyways, this PR updates us to the latest version of NUnit and
hopes it fixes the issue.  🤷‍♂️

[0]: https://github.com/nunit/nunit3-vs-adapter/issues/649