[XABT] Do not prefer NDKMinimumApiAvailable
I noticed that we were not writing a valid `minSdkVersion` value to the
generated `AndroidManifest.xml` file for projects which included a
manifest that declared a `targetSdkVersion`.  In these cases, we would
always write a `minSdkVersion` of `19` to the manifest file, as this is
the min version that our NDK supports.

Fix this to always write the value of `$(SupportedOSPlatformVersion)`.
If this value is not explicitly set in the project file, it will default
to `$(TargetPlatformVersion)`.

A couple of multidex related tests have been added to the ignore list as
they are only valid with a minSdkVersion of 20 or lower.