Way too much time spent on this...
...but LLVM IR conditional blocks finally work and IR quirks worked
around.

The main quirk is that in LLVM IR all unnamed parameters, local
variables **and** labels must use sequential numbers as their name.
However, a branch instruction needs to be (obviously) generated before
the labels it refers to (two of them, for true and false conditions) are
output.  This poses a problem if code generation is done in one,
forward-only, pass - like in our case.  There's no way to go back to the
br instruction and update label names (which would be easy if an AST was
used), so instead of using anonymous labels we now use named ones but
whose names have a sequential integer value appended to guarantee
uniqueness should two labels of the same name be requested.