[Xamarin.Android.Build.Tasks] introduce Eol.targets for .NET 6 (#8047)
Fixes: https://github.com/xamarin/xamarin-android/issues/8003

Add a new `Eol.targets` file to be imported for `net6.0-android`
projects.  It contains the necessary/minimal settings to emit the
build error:

	Microsoft.NET.EolTargetFrameworks.targets(35,5): error NETSDK1202:
	The workload 'android' is out of support and will not receive security updates in the future.
	Please refer to https://aka.ms/maui-support-policy for more information about the support policy.

We will probably need the latest .NET SDK to flow to us, in order to
write a test.  I could only test locally by editing `.targets` files
in the .NET SDK.