[XABT] Prefer `SupportedOSPlatformVersion` over `minSdkVersion` (#8026)
Fixes: https://github.com/xamarin/xamarin-android/issues/8040

I noticed that we were not writing a supported (by us)
`//uses-sdk/@android:minSdkVersion` value to the generated
`AndroidManifest.xml` file for projects which included a manifest
that declared a `targetSdkVersion`.  In these cases, we would
always write a `//uses-sdk/@android:minSdkVersion` value of `19` to
`AndroidManifest.xml`, as that was the minimum API level the
Android NDK supported.  (This was also a mistake, as *MonoVM* has
API-21 as the minimum supported target.  This mistake is corrected.)

Fix this to always use the value of `$(SupportedOSPlatformVersion)`
as the `minSdkVersion` attribute in the `AndroidManifest.xml` file.
If this value is not explicitly set in the project file, it will now
default to `$(AndroidMinimumSupportedApiLevel)` instead of
`$(TargetPlatformVersion)`.

The `XA4216` error/warning code has been expanded+repurposed.

The warning that would display when the `minSdkVersion` attribute in
`AndroidManifest.xml` was less than our minimum supported API level
has been converted into an error:

	error XA4216: The deployment target '19' is not supported (the minimum is '21').
	Please increase (or remove) the //uses-sdk/@android:minSdkVersion value in your project file.

A similar error is now reported when `$(SupportedOSPlatformVersion)`
is less than `$(AndroidMinimumSupportedApiLevel)`

	error XA4216: The deployment target '19' is not supported (the minimum is '21').
	Please increase the $(SupportedOSPlatformVersion) property value in your project file.

An error condition has been added for when
`//uses-sdk/@android:minSdkVersion` in `AndroidManifest.xml` does not
match `$(SupportedOSPlatformVersion)`:

	error XA1036: AndroidManifest.xml //uses-sdk/@android:minSdkVersion '19' does not match the $(SupportedOSPlatformVersion) value '21.0' in the project file (if there is no $(SupportedOSPlatformVersion) value in the project file, then a default value has been assumed).
	Either change the value in the AndroidManifest.xml to match the $(SupportedOSPlatformVersion) value, or remove the value in the AndroidManifest.xml (and add a $(SupportedOSPlatformVersion) value to the project file if it doesn't already exist).

A few multidex related tests have been updated/removed as they are
only valid with a `minSdkVersion` of 20 or lower.