[Xamarin.Android.Build.Tasks] Ignore non-Android XML resources (#8091)
We have seen reports of the following error/warning when building
Maui or Android projects.

	error XAGRDA7005: System.FormatException: Could not find any recognizable digits.  
	   at System.ParseNumbers.StringToInt(String s, Int32 radix, Int32 flags, Int32* currPos)   
	   at System.Convert.ToInt32(String value, Int32 fromBase)   
	   at Xamarin.Android.Tasks.RtxtParser.ProcessRtxtFile(String file, IList`1 result)   
	   at Xamarin.Android.Tasks.RtxtParser.Parse(String file, TaskLoggingHelper logger, Dictionary`2 mapping)   
	   at Xamarin.Android.Tasks.GenerateResourceDesignerAssembly.Run(DirectoryAssemblyResolver res)   
	   at Xamarin.Android.Tasks.GenerateResourceDesignerAssembly.RunTask()   
	   at Microsoft.Android.Build.Tasks.AndroidTask.Execute() in /Users/runner/work/1/s/xamarin-android/external/xamarin-android-tools/src/Microsoft.Android.Build.BaseTasks/AndroidTask.cs:line 25

Our initial thoughts were that we were trying to parse an empty
`R.txt` file.  However this did not seem to pan out.  Further
investigation found this line in an `R.txt` file

	int id maximum number 0x7f037465

This is very odd, since `id` values should NOT contain spaces.
It turns out this is caused by the following resource fragment:

	<resources xmlns:ns1="http://schemas.android.com/tools" xmlns:ns2="urn:oasis:names:tc:xliff:document:1.2">
	  <string name="translatable_text" translatable="false">
	    <ns2:g example="999" id="maximum number">%1$d</ns2:g>
	    <ns2:g example="+" id="suffix">%2$s</ns2:g>
	  </string>
	</resources>

This is from `com.google.android.material.material.aar`.  It turns
out that you can have additional namespaces in the resource xml files.
`aapt2` ignores these additional namespaces but OUR managed parser
does not.   As a result the `id="maximum number"` attribute gets
picked up as a valid `id` value.

The fix in this case is to check if the containing element has a
namespace.  If it does then make sure it is the android resource
namespace.  ALL other namespaces will be ignored.  We also added a
new Regex to check for invalid characters just incase the Namespace
check allows some entires though.